EPMDJ-5962: Write tests for graph calculation functions.
Issue: EPMDJ-5962
