+++ src/modules/tests-to-run/tests-to-run-list/bar-chart/bar-chart.tsx
@@ -22,6 +22,10 @@ import { DATA_VISUALIZATION_COLORS } from 'common/constants';
 import { useBuildVersion, useElementSize } from 'hooks';
 import { TestsToRunSummary } from 'types/tests-to-run-summary';
 import { getDuration, percentFormatter } from 'utils';
+import { getTimeFormat, getYScale } from './helpers';
+import {
+  BAR_HORIZONTAL_PADDING_PX, BAR_WITH_GAP_WIDTH_PX, CHART_HEIGHT_PX, BORDER_PX,
+} from './constants';
 
 import styles from './bar-chart.module.scss';
 
@@ -33,11 +37,6 @@ interface Props {
 }
 
 const barChart = BEM(styles);
-
-const BAR_HORIZONTAL_PADDING_PX = 96;
-const BAR_WITH_GAP_WIDTH_PX = 112;
-const CHART_HEIGHT_PX = 160;
-const BORDER_PX = 1;
 
 export const BarChart = barChart(({
   className, activeBuildVersion, totalDuration, summaryTestsToRun,
@@ -192,192 +191,3 @@ const YAxis = barChart.yAxis('div');
 const XAxis = barChart.xAxis('div');
 const XAxisLegend = barChart.xAxisLegend('div');
 const SavedTimePercent = barChart.savedTimePercent('div');
-
-const MS_IN_SECONDS = 1000;
-const MS_IN_MINUTES = MS_IN_SECONDS * 60;
-const MS_IN_HOURS = MS_IN_MINUTES * 60;
-
-const graphParams = [
-  {
-    duration: 100 * MS_IN_HOURS,
-    step: 25 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 90 * MS_IN_HOURS,
-    step: 30 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 60 * MS_IN_HOURS,
-    step: 20 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 45 * MS_IN_HOURS,
-    step: 15 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 30 * MS_IN_HOURS,
-    step: 10 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 15 * MS_IN_HOURS,
-    step: 5 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 12 * MS_IN_HOURS,
-    step: 3 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 6 * MS_IN_HOURS,
-    step: 2 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 3 * MS_IN_HOURS,
-    step: 1 * MS_IN_HOURS,
-    unit: 'h',
-  },
-  {
-    duration: 1 * MS_IN_HOURS,
-    step: 20 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  // duration < 1 hour
-  {
-    duration: 45 * MS_IN_MINUTES,
-    step: 15 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  {
-    duration: 30 * MS_IN_MINUTES,
-    step: 10 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  {
-    duration: 15 * MS_IN_MINUTES,
-    step: 5 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  {
-    duration: 12 * MS_IN_MINUTES,
-    step: 3 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  {
-    duration: 6 * MS_IN_MINUTES,
-    step: 2 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  {
-    duration: 3 * MS_IN_MINUTES,
-    step: 1 * MS_IN_MINUTES,
-    unit: 'm',
-  },
-  {
-    duration: 1 * MS_IN_MINUTES,
-    step: 20 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  // duration < 1m
-  {
-    duration: 45 * MS_IN_SECONDS,
-    step: 15 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  {
-    duration: 30 * MS_IN_SECONDS,
-    step: 10 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  {
-    duration: 15 * MS_IN_SECONDS,
-    step: 5 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  {
-    duration: 12 * MS_IN_SECONDS,
-    step: 3 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  {
-    duration: 6 * MS_IN_SECONDS,
-    step: 2 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  {
-    duration: 3 * MS_IN_SECONDS,
-    step: 1 * MS_IN_SECONDS,
-    unit: 's',
-  },
-  {
-    duration: 1 * MS_IN_SECONDS,
-    step: 200,
-    unit: 'ms',
-  },
-];
-
-// TODO move that
-function getTimeFormat(durationMs: number, unit: string) {
-  switch (unit) {
-    case 'h':
-      return durationMs / 3600000;
-    case 'm':
-      return durationMs / 60000;
-    case 's':
-      return durationMs / 1000;
-    default:
-      return durationMs;
-  }
-}
-
-function getYScale(duration: number) {
-  const params = getParams(duration);
-  const count = Math.ceil(duration / params.step);
-  return {
-    stepSizeMs: params.step,
-    unit: params.unit,
-    stepsCount: count,
-  };
-}
-
-function getParams(duration: number) {
-  const durationsMs = graphParams.map(x => x.duration);
-
-  const lesserNeighborIndex = durationsMs.findIndex(x => x <= duration);
-
-  const isOutOfRange = lesserNeighborIndex === -1;
-  if (isOutOfRange) {
-    const isLessThanRange = duration <= durationsMs[durationsMs.length - 1];
-    if (isLessThanRange) {
-      return graphParams[graphParams.length - 1];
-    }
-    return graphParams[0];
-  }
-
-  const exactGt = lesserNeighborIndex === durationsMs.length - 1;
-  if (exactGt) {
-    return graphParams[graphParams.length - 1];
-  }
-
-  const exactLt = lesserNeighborIndex === 0;
-  if (exactLt) {
-    return graphParams[0];
-  }
-
-  const lesserError = Math.abs(durationsMs[lesserNeighborIndex] - duration);
-  const greaterError = Math.abs(durationsMs[lesserNeighborIndex - 1] - duration);
-
-  const bestNeighborIndex = lesserError < greaterError
-    ? lesserNeighborIndex
-    : lesserNeighborIndex - 1;
-
-  const step = graphParams[bestNeighborIndex];
-
-  return step;
-}
