+++ src/modules/tests-to-run/tests-to-run-list/bar-chart/helpers/get-time-format.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2020 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+type Unit = 'h' | 'm' | 's';
+
+export function getTimeFormat(durationMs: number, unit: Unit) {
+  if (Number.isNaN(durationMs) || !Number.isFinite(durationMs)) return 0;
+  switch (unit) {
+    case 'h':
+      return durationMs / 3600000;
+    case 'm':
+      return durationMs / 60000;
+    case 's':
+      return durationMs / 1000;
+    default:
+      return durationMs;
+  }
+}
